[#30][REFACTOR] extract shared agent fixtures to conftest.py and fix lint
